fix: add temporary fix
we reroute from the login page to the main page when the user logs in. that unmounts the login component and cancels the generateNewToken call (in the apollo client middleware). as a temporary fix, we add a pseud sleep function to delay the dismount and allow the generateNewToken to complete